requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==2.2.4
opencv-python==4.11.0.86
pillow==11.2.1
azure-storage-blob==12.25.1
